+++ CHANGES.md
@@ -36,6 +36,8 @@ To be released.
 
 ### Added interfaces
 
+ -  Added `ProtectedPrivateKey` class.  [[#577], [#614]]
+ -  Added `IncorrectPassphraseException` class.  [[#577], [#614]]
  -  Added `ICipher` interface.  [[#577], [#614]]
  -  Added `Aes128Ctr` class.  [[#577], [#614]]
  -  Added `IKdf` interface.  [[#577], [#614]]